+++ service/state/index.go
@@ -19,12 +19,12 @@ import (
 	"fmt"
 
 	"github.com/OneOfOne/xxhash"
-	"github.com/awfm9/flow-dps/service/state/storage"
 	"github.com/dgraph-io/badger/v2"
 
 	"github.com/onflow/flow-go/model/flow"
 
 	"github.com/awfm9/flow-dps/models/dps"
+	"github.com/awfm9/flow-dps/service/storage"
 )
 
 type Index struct {
